requirements/dev.txt
wagtail==2.7.1
dj-database-url==0.5.0
Django==2.2.4
django-configurations==2.1
psycopg2-binary==2.8.3
python_dotenv==0.10.3
markdown2==2.3.8
Pillow==6.1.0
django-extensions==2.2.1
martor==1.4.1
gunicorn==19.9.0
docutils==0.15.2
pytest==5.1.2
pytest-cov==2.7.1
pytest-django==3.5.1
selenium==3.141.0
docutils==0.15.2
flake8==3.7.8
pylint==2.3.1
Werkzeug==0.15.5
pyOpenSSL==19.0.0
ipython==7.7.0
cryptography==2.7
patchwork==1.0.1
sentry-sdk==0.12.2
